refactor(kftray-kube): promote http_log_watcher to crate root
