+++ crates/kftray-kube/src/http_log_watcher.rs
